cr-354: Refactor get_codes package to avoid xml import in csv writer
