+++ src/pysdmx/toolkit/pd/_data_utils.py
@@ -1,8 +1,8 @@
-from typing import Literal, Sequence
+from typing import Any, Dict, List, Literal, Sequence, Tuple
 
 import pandas as pd
 
-from pysdmx.model.dataflow import Component
+from pysdmx.model.dataflow import Component, Schema
 
 
 def format_labels(  # noqa: C901
@@ -47,3 +47,54 @@ def format_labels(  # noqa: C901
                         columns={k: component.concept.id},
                         inplace=True,
                     )
+
+
+def get_codes(
+    dimension_code: str, structure: Schema, data: pd.DataFrame
+) -> Tuple[List[str], List[str], List[Dict[str, Any]]]:
+    """This function divides the components in Series and Obs."""
+    series_codes = []
+    groups = structure.groups
+    group_codes = []
+    obs_codes = [dimension_code, structure.components.measures[0].id]
+
+    # Getting the series and obs codes
+    for dim in structure.components.dimensions:
+        if dim.id != dimension_code:
+            series_codes.append(dim.id)
+
+    # Adding the attributes based on the attachment level
+    for att in structure.components.attributes:
+        matching_group = next(
+            (
+                group
+                for group in groups or []
+                if set(group.dimensions)
+                == set(att.attachment_level.split(","))  # type: ignore[union-attr]
+            ),
+            None,
+        )
+
+        if (
+            att.attachment_level != "D"
+            and att.id in data.columns
+            and groups is not None
+            and matching_group
+        ):
+            group_codes.append(
+                {
+                    "group_id": matching_group.id,
+                    "attribute": att.id,
+                    "dimensions": matching_group.dimensions,
+                }
+            )
+        elif att.attachment_level == "O" and att.id in data.columns:
+            obs_codes.append(att.id)
+        elif (
+            att.attachment_level is not None
+            and att.attachment_level != "D"
+            and att.id in data.columns
+        ):
+            series_codes.append(att.id)
+
+    return series_codes, obs_codes, group_codes
